Safari-611.1.10/LayoutTests/accessibility/mac/submit-button-default-value-expected.txt


This tests that a submit button will return its default value as the title.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.title is 'AXTitle: Submit'
PASS button.title is 'AXTitle: non-default value'
PASS successfullyParsed is true

TEST COMPLETE
